Fix: use ms instead of s for got Retry-After limit
`got`'s documentation is incorrect/unclear and looking through the source code, we can see that the library actually expects `maxRetryAfter` to be milliseconds instead of seconds like used in the `Retry-After` header (of which docs are linked in the `maxRetryAfter` input documentation...). This being in seconds, led to the behaviour that basically none of the requests were getting retried as the limit was always lower than the lowest `Retry-After` value (1 second). This, combined with the fact that the errors don't specifically mention the limit being hit, meant that we couldn't identify the real issue immediately.